Cast to Make Sure Proper Overloaded Method Called

+++ examples/RPi_4/bcm2835/example_rpi4_bcm2835.cpp
@@ -17,7 +17,7 @@ int main()
 {
   I2C_Interface* i2c_if = new BCM2835_I2C_IF();
   /* initialize fast I2C 400 kHz */
-  if(i2c_if->Init_I2C(400000) != I2C_STATUS_SUCCESS)
+  if(i2c_if->Init_I2C((uint32_t)400000) != I2C_STATUS_SUCCESS)
   {
     std::cout << "I2C initialization failed!\n";
     delete i2c_if;
